move RequestController up
