--- src/core/controller/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  RequestController: require("./RequestController")
-};
